% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio_cons_est.R
\name{bio_cons_est}
\alias{bio_cons_est}
\title{Estimating energyscape and biomass consumption}
\usage{
bio_cons_est(
  predator_group,
  predator_name,
  beta = NULL,
  prey_taxonomic_level,
  prey_group,
  diet,
  assimilation_rate = 0.8,
  n_days = 1,
  predator_weight,
  abundance_map,
  temperature_map = NULL
)
}
\arguments{
\item{predator_group}{Character. Predator taxonomic group, used to select the allometric equation used to compute the analysis. Can be "Cetacean", "Fish", "Charadriform", "Procellariform", "Loggerhead turtle"}

\item{predator_name}{Character. Name to be used in exported dataframes}

\item{beta}{Vector of life cost parameter, used only for mammals}

\item{prey_taxonomic_level}{Taxonomic level at which the diet is summarize (either "Family", "Taxonomic_group" or "Functional_group", for example), i.e. the column in which prey_group will be subsetted}

\item{prey_group}{Which level of the diet to select (one of the levels in the chosen prey_taxonomic_level)}

\item{diet}{Diet table, including the considered predator only. Must include pW (proportion of each item in the diet, as proportion of wet weight), Energy_content of the prey item (in kJ/g) and the column provided in prey_taxonomic_group}

\item{assimilation_rate}{Assimilation rate to be used to estimate the daily ration (default to 0.8)}

\item{n_days}{Number of days over which estimating biomas consumption (default to 1)}

\item{predator_weight}{Vector of body mass distribution for the predator (in kg)}

\item{abundance_map}{Abundance map of the predator (must provide columns named mean and sd)}

\item{temperature_map}{SST map (only used for fish; must provide columns named mean and sd)}
}
\value{
The function returns a list composed of 6 elements: FMR, the energetic needs of a single individuals (kJ/d); abundance, the abubndance map; Energyscape, the energyscape map for the total predator population, for a single day (FMR*abundance; kJ/d); daily_ration, the estimated daily ration (in kg); prop_body_mass, the daily ration expressed as proportion of predator body mass; conso_map, a map of the prey_group biomass consumed by the predator population over the period provided by n_days.
}
\description{
The function first estimates the energyscape (spatialised energetic need of the population) of a marine predator species (either cetacean, seabirds, loggerhead turtles, sunfish, sharks, tunas or swordfish) from its abundance map and body mass distribution, as well as the sea surface temperature (for fish only). Second, the function estimate the daily ration consumed for a given prey group by a single predator individual and the corresponding biomass consumed by the total population. Results are provided as maps and vectors. This function runs a single estimation.
Allometric relationships are derived from White et al. (2011) with beta parameters from Spitz et al. (2012) for cetaceans; from Ultsch et al. 2013 for loggerhead turtle; from Killen et al. (2016) for fishes and from Shaffer et al. (2011) for seabirds (contrasting Procellariforms and Charadriforms).
}
\examples{
result <- bio_cons_est(predator_group = "Cetacean",
             predator_name = "TURTRU", 
             prey_taxonomic_level = "Taxonomic_group", 
             prey_group = "Cephalopoda",
             diet = diet, 
             assimilation_rate = 0.8,
             n_days = 30+31+31,
             predator_weight = weight, 
             abundance_map = species_abundance, 
             temperature_map = sst,
             beta = beta) # not used for fishes and seabirds

}
