% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_diet.R
\docType{data}
\name{diet}
\alias{diet}
\title{diet}
\format{
A data frame with 30 rows and 14 variables:
\describe{
\item{ Predator_key }{  character }
\item{ Latin_name }{  character }
\item{ pW }{  numeric }
\item{ pW_uncertainty }{  logical }
\item{ Source.x }{  character }
\item{ Common_name }{  character }
\item{ Energy_content }{  numeric }
\item{ Energy_content_uncertainty }{  logical }
\item{ Estimate_level }{  character }
\item{ Family }{  character }
\item{ Taxonomic_group }{  character }
\item{ Functional_group }{  character }
\item{ Source.y }{  character }
\item{ ...10 }{  character }
}
}
\source{
Source
}
\usage{
diet
}
\description{
Description.
}
\keyword{datasets}
