% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoscope_pal.R
\name{ecoscope_pal}
\alias{ecoscope_pal}
\title{Function to interpolate a color palette}
\usage{
ecoscope_pal(palette = "ecoscope1", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in ecoscope_palette}

\item{reverse}{Boolean true if palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Function to interpolate a color palette
}
\examples{
ggplot2::ggplot(diet) +
  ggplot2::geom_col(ggplot2::aes(x = Predator_key, y = pW, fill = Functional_group)) +
  scale_fill_ecoscope() 
}
