% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoscope_pal.R
\docType{data}
\name{ecoscope_palette}
\alias{ecoscope_palette}
\title{Color palettes}
\format{
An object of class \code{list} of length 1.
}
\usage{
ecoscope_palette
}
\description{
Color palettes
}
\keyword{datasets}
