% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energyscapes-package.R
\docType{package}
\name{energyscapes-package}
\alias{energyscapes}
\alias{energyscapes-package}
\title{energyscapes: Toolbox For Energyscape Analyses}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Estimate energyscape and biomass consumption of top predators from allometric relationships.
}
\author{
\strong{Maintainer}: Charlotte Lambert \email{charlotte.lambert@univ-lr.fr} (\href{https://orcid.org/0000-0002-1128-5262}{ORCID})

}
\keyword{internal}
