% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_map_coords.R
\docType{data}
\name{map_coords}
\alias{map_coords}
\title{map_coords}
\format{
A data frame with 5720 rows and 2 variables:
\describe{
\item{ x }{  numeric }
\item{ y }{  numeric }
}
}
\source{
Source
}
\usage{
map_coords
}
\description{
Coordinates of the map.
}
\keyword{datasets}
