% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoscope_pal.R
\name{scale_fill_ecoscope}
\alias{scale_fill_ecoscope}
\title{Fill scale for Ecoscope colors}
\usage{
scale_fill_ecoscope(
  palette = "ecoscope1",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in ecoscope_palette}

\item{discrete}{Boolean if color aesthetic is discrete}

\item{reverse}{Boolean if palette should be reversed}

\item{...}{Additional arguments used to discrete_scale() or scale_fill_gradientn()
to automatically interpolate between colours.}
}
\description{
Fill scale for Ecoscope colors
}
