% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_species_abundance.R
\docType{data}
\name{species_abundance}
\alias{species_abundance}
\title{species_abundance}
\format{
A data frame with 5720 rows and 4 variables:
\describe{
\item{ x }{  x coordinates }
\item{ y }{  y coordinates }
\item{ mean }{  mean number of individuals per cell }
\item{ sd }{  standard deviation of the number of individuals per cell }
\item{ bathy }{ bathymetry in each cell }
}
}
\source{
Source
}
\usage{
species_abundance
}
\description{
Abundance maps of Scopoli's shearwater, from Lambert et al 2024 ICES JMS (10.1093/icesjms/fsae058).
}
\keyword{datasets}
