% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_sst.R
\docType{data}
\name{sst}
\alias{sst}
\title{sst}
\format{
A data frame with 5720 rows and 4 variables:
\describe{
\item{ x }{  numeric }
\item{ y }{  numeric }
\item{ mean }{  numeric }
\item{ sd }{  numeric }
}
}
\source{
Source
}
\usage{
sst
}
\description{
Description.
}
\keyword{datasets}
