% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_weight.R
\docType{data}
\name{weight}
\alias{weight}
\title{weight}
\format{
A vector of body mass distribution (kg):
\describe{
}
}
\source{
Source
}
\usage{
weight
}
\description{
Weight distribution in Scopoli's shearwater populations.
}
\keyword{datasets}
