% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_species_abundance.R
\docType{data}
\name{species_abundance}
\alias{species_abundance}
\title{species_abundance}
\format{
A data frame with 5720 rows and 4 variables:
\describe{
\item{ x }{  numeric }
\item{ y }{  numeric }
\item{ mean }{  numeric }
\item{ sd }{  numeric }
}
}
\source{
Source
}
\usage{
species_abundance
}
\description{
Description.
}
\keyword{datasets}
