% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_sst.R
\docType{data}
\name{sst}
\alias{sst}
\title{sst}
\format{
A data frame with 5720 rows and 4 variables:
\describe{
\item{ x }{  x coordinates }
\item{ y }{  y coordinates }
\item{ mean }{  mean SST }
\item{ sd }{  standard deviation of SST }
}
}
\source{
Source
}
\usage{
sst
}
\description{
Sea surface temperature from Copernicus Marine Service product "SST MEDSEA MULTIYEAR PHY 006 004" from https://doi.org/10.25423/CMCC/MEDSEA MULTIYEAR PHY 006 004 E3R1.
}
\keyword{datasets}
